+++ src/main/api/ledger/address.ts
@@ -1,9 +1,10 @@
 import TransportNodeHidSingleton from '@ledgerhq/hw-transport-node-hid-singleton'
 import { Address } from '@xchainjs/xchain-client'
-import { THORChain } from '@xchainjs/xchain-util'
+import { BNBChain, THORChain } from '@xchainjs/xchain-util'
 import * as E from 'fp-ts/Either'
 
 import { IPCLedgerAdddressParams, LedgerErrorId } from '../../../shared/api/types'
+import { getAddress as getBNBAddress } from './binance'
 import { getAddress as getTHORAddress } from './thorchain/address'
 import { getErrorId } from './utils'
 
@@ -17,6 +18,9 @@ export const getAddress = async ({
     switch (chain) {
       case THORChain:
         res = await getTHORAddress(transport, network)
+        break
+      case BNBChain:
+        res = await getBNBAddress(transport, network)
         break
       default:
         res = E.left(LedgerErrorId.NO_APP)