+++ src/renderer/components/pool/PoolTitle.tsx
@@ -24,6 +24,7 @@ export type Props = {
   isLoading?: boolean
   disableTradingPoolAction: boolean
   disableAllPoolActions: boolean
+  disablePoolActions: boolean
   network: Network
   status: GetPoolsStatusEnum
 }
@@ -34,6 +35,7 @@ export const PoolTitle: React.FC<Props> = ({
   priceSymbol,
   disableTradingPoolAction,
   disableAllPoolActions,
+  disablePoolActions,
   network,
   isLoading,
   status
@@ -102,7 +104,7 @@ export const PoolTitle: React.FC<Props> = ({
             return (
               <Styled.ButtonActions>
                 <ManageButton
-                  disabled={disableAllPoolActions}
+                  disabled={disableAllPoolActions || disablePoolActions}
                   asset={asset}
                   sizevalue={isDesktopView ? 'normal' : 'small'}
                   isTextView={isDesktopView}
@@ -129,7 +131,7 @@ export const PoolTitle: React.FC<Props> = ({
           }
         )
       ),
-    [oAsset, disableAllPoolActions, isDesktopView, status, disableTradingPoolAction, intl, history]
+    [oAsset, disableAllPoolActions, disablePoolActions, isDesktopView, status, disableTradingPoolAction, intl, history]
   )
 
   return (