+++ src/renderer/helpers/poolHelper.test.ts
@@ -9,7 +9,9 @@ import {
   AssetBNB,
   ETHChain,
   BNBChain,
-  LTCChain
+  LTCChain,
+  BTCChain,
+  BCHChain
 } from '@xchainjs/xchain-util'
 import * as FP from 'fp-ts/lib/function'
 import * as O from 'fp-ts/lib/Option'
@@ -21,6 +23,7 @@ import { DEFAULT_MIMIR_HALT } from '../services/thorchain/const'
 import { GetPoolsStatusEnum, PoolDetail } from '../types/generated/midgard'
 import {
   disableAllActions,
+  disablePoolActions,
   disableTradingActions,
   getDeepestPool,
   getPoolPriceValue,
@@ -238,7 +241,10 @@ describe('helpers/poolHelper/', () => {
       const result = disableTradingActions({
         chain: BNBChain,
         haltedChains,
-        mimirHalt: { haltTrading: true, haltEthTrading: false }
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT,
+          haltTrading: true
+        }
       })
       expect(result).toBeTruthy()
     })
@@ -246,7 +252,21 @@ describe('helpers/poolHelper/', () => {
       const result = disableTradingActions({
         chain: LTCChain,
         haltedChains,
-        mimirHalt: { haltTrading: true, haltEthTrading: false }
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT,
+          haltTrading: true
+        }
+      })
+      expect(result).toBeTruthy()
+    })
+    it('true for BTC if BTC trading is halted', () => {
+      const result = disableTradingActions({
+        chain: BTCChain,
+        haltedChains,
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT,
+          haltBtcTrading: true
+        }
       })
       expect(result).toBeTruthy()
     })
@@ -254,15 +274,57 @@ describe('helpers/poolHelper/', () => {
       const result = disableTradingActions({
         chain: ETHChain,
         haltedChains,
-        mimirHalt: { haltTrading: false, haltEthTrading: true }
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT,
+          haltEthTrading: true
+        }
       })
       expect(result).toBeTruthy()
     })
-    it('false for a chain, if it is not in halted list, but ETH trading is halted', () => {
+    it('true for BCH if BCH trading is halted', () => {
+      const result = disableTradingActions({
+        chain: BCHChain,
+        haltedChains,
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT,
+          haltBchTrading: true
+        }
+      })
+      expect(result).toBeTruthy()
+    })
+    it('true for LTC if LTC trading is halted', () => {
       const result = disableTradingActions({
         chain: LTCChain,
         haltedChains,
-        mimirHalt: { haltTrading: false, haltEthTrading: true }
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT,
+          haltLtcTrading: true
+        }
+      })
+      expect(result).toBeTruthy()
+    })
+    it('true for BNB if BNB trading is halted', () => {
+      const result = disableTradingActions({
+        chain: BNBChain,
+        haltedChains,
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT,
+          haltBnbTrading: true
+        }
+      })
+      expect(result).toBeTruthy()
+    })
+    it('false for a chain, if it is not in halted list, but other chains have trading halted', () => {
+      const result = disableTradingActions({
+        chain: LTCChain,
+        haltedChains,
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT,
+          haltBtcTrading: true,
+          haltEthTrading: true,
+          haltBchTrading: true,
+          haltBnbTrading: true
+        }
       })
       expect(result).toBeFalsy()
     })
@@ -270,7 +332,9 @@ describe('helpers/poolHelper/', () => {
       const result = disableTradingActions({
         chain: ETHChain,
         haltedChains,
-        mimirHalt: { haltTrading: false, haltEthTrading: false }
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT
+        }
       })
       expect(result).toBeTruthy()
     })
@@ -278,7 +342,9 @@ describe('helpers/poolHelper/', () => {
       const result = disableTradingActions({
         chain: BNBChain,
         haltedChains,
-        mimirHalt: { haltTrading: false, haltEthTrading: false }
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT
+        }
       })
       expect(result).toBeTruthy()
     })
@@ -286,9 +352,63 @@ describe('helpers/poolHelper/', () => {
       const result = disableTradingActions({
         chain: LTCChain,
         haltedChains,
-        mimirHalt: { haltTrading: false, haltEthTrading: false }
+        mimirHalt: {
+          ...DEFAULT_MIMIR_HALT,
+          haltEthTrading: true
+        }
       })
       expect(result).toBeFalsy()
+    })
+  })
+
+  describe('disablePoolActions', () => {
+    it('true if trading is halted for this chain', () => {
+      const result = disablePoolActions({
+        chain: BNBChain,
+        haltedChains: [ETHChain, BNBChain],
+        mimirHalt: { ...DEFAULT_MIMIR_HALT }
+      })
+      expect(result).toBeTruthy()
+    })
+    it('true if BNB chain is not in halted list, but paused', () => {
+      const result = disablePoolActions({
+        chain: BNBChain,
+        haltedChains: [],
+        mimirHalt: { ...DEFAULT_MIMIR_HALT, pauseLpBnb: true }
+      })
+      expect(result).toBeTruthy()
+    })
+    it('true if BTC chain is not in halted list, but paused', () => {
+      const result = disablePoolActions({
+        chain: BTCChain,
+        haltedChains: [],
+        mimirHalt: { ...DEFAULT_MIMIR_HALT, pauseLpBtc: true }
+      })
+      expect(result).toBeTruthy()
+    })
+    it('true if BCH chain is not in halted list, but paused', () => {
+      const result = disablePoolActions({
+        chain: BCHChain,
+        haltedChains: [],
+        mimirHalt: { ...DEFAULT_MIMIR_HALT, pauseLpBch: true }
+      })
+      expect(result).toBeTruthy()
+    })
+    it('true if ETH chain is not in halted list, but paused', () => {
+      const result = disablePoolActions({
+        chain: ETHChain,
+        haltedChains: [],
+        mimirHalt: { ...DEFAULT_MIMIR_HALT, pauseLpEth: true }
+      })
+      expect(result).toBeTruthy()
+    })
+    it('true if LTC chain is not in halted list, but paused', () => {
+      const result = disablePoolActions({
+        chain: LTCChain,
+        haltedChains: [],
+        mimirHalt: { ...DEFAULT_MIMIR_HALT, pauseLpLtc: true }
+      })
+      expect(result).toBeTruthy()
     })
   })
 })