+++ src/renderer/helpers/poolHelper.ts
@@ -12,7 +12,7 @@ import { Network } from '../../shared/api/types'
 import { ONE_RUNE_BASE_AMOUNT } from '../../shared/mock/amount'
 import { PoolAddress, PoolDetails } from '../services/midgard/types'
 import { getPoolDetail, toPoolData } from '../services/midgard/utils'
-import { MimirHalt, MimirHaltChain } from '../services/thorchain/types'
+import { MimirHaltChain, MimirHaltTrading, MimirPauseLP } from '../services/thorchain/types'
 import { PoolDetail } from '../types/generated/midgard'
 import { PoolTableRowData, PoolTableRowsData, PricePool } from '../views/pools/Pools.types'
 import { getPoolTableRowData } from '../views/pools/Pools.utils'
@@ -165,6 +165,12 @@ const isChainElem = A.elem(eqChain)
 
 /**
  * Helper to check if all pool actions (`SWAP`, `ADD`, `WITHDRAW`) have to be disabled
+ *
+ * |                  | ADD | WITHDRAW | SWAP |
+ * |------------------|-----|----------|------|
+ * | halt{chain}Chain | NO  | NO       | NO   |
+ * | halt{chain}      | NO  | NO       | NO   |
+ *
  */
 export const disableAllActions = ({
   chain,
@@ -199,22 +205,59 @@ export const disableAllActions = ({
 
 /**
  * Helper to check if pool trading actions (`SWAP`, `ADD`) have to be disabled
+ *
+ * |                    | ADD | WITHDRAW | SWAP |
+ * |--------------------|-----|----------|------|
+ * | halt{chain}Trading | NO  | YES      | NO   |
+ * | halt{chain}        | NO  | NO       | NO   |
  */
 export const disableTradingActions = ({
   chain,
   haltedChains,
-  mimirHalt: { haltTrading, haltEthTrading }
+  mimirHalt: { haltTrading, haltBtcTrading, haltEthTrading, haltBchTrading, haltLtcTrading, haltBnbTrading }
 }: {
   chain: Chain
   haltedChains: Chain[]
-  mimirHalt: Pick<MimirHalt, 'haltTrading' | 'haltEthTrading'>
+  mimirHalt: MimirHaltTrading
 }) => {
   // 1. Check `haltTrading` (provided by `mimir` endpoint) to disable all actions for all pools
   if (haltTrading) return true
 
-  // 2. Check `haltEthTrading` (provided by `mimir` endpoint) to disable all actions for ETH pools
+  // 2. Check if trading is disabled for chain to disable all actions for pool
+  if (isBtcChain(chain) && haltBtcTrading) return true
   if (isEthChain(chain) && haltEthTrading) return true
+  if (isBchChain(chain) && haltBchTrading) return true
+  if (isLtcChain(chain) && haltLtcTrading) return true
+  if (isBnbChain(chain) && haltBnbTrading) return true
 
   // 3. Check `chain` is included in `haltedChains` (provided by `inbound_addresses` endpoint)
+  return FP.pipe(haltedChains, isChainElem(chain))
+}
+
+/**
+ * Helper to check if pool trading actions (`ADD`, `WITHDRAW`) have to be disabled
+ *
+ * |                | ADD | WITHDRAW | SWAP |
+ * |----------------|-----|----------|------|
+ * | pauseLP{chain} | NO  | NO       | YES  |
+ * | halt{chain}    | NO  | NO       | NO   |
+ */
+export const disablePoolActions = ({
+  chain,
+  haltedChains,
+  mimirHalt: { pauseLpBnb, pauseLpBch, pauseLpBtc, pauseLpEth, pauseLpLtc }
+}: {
+  chain: Chain
+  haltedChains: Chain[]
+  mimirHalt: MimirPauseLP
+}) => {
+  // Check all `pauseLp{chain}` values (provided by `mimir` endpoint) to disable pool actions
+  if (isBnbChain(chain) && pauseLpBnb) return true
+  if (isBchChain(chain) && pauseLpBch) return true
+  if (isBtcChain(chain) && pauseLpBtc) return true
+  if (isEthChain(chain) && pauseLpEth) return true
+  if (isLtcChain(chain) && pauseLpLtc) return true
+
+  // Check `chain` is included in `haltedChains` (provided by `inbound_addresses` endpoint)
   return FP.pipe(haltedChains, isChainElem(chain))
 }