+++ src/renderer/i18n/de/pools.ts
@@ -8,8 +8,7 @@ const pools: PoolsMessages = {
   'pools.blocksleft': 'Verbleibende Blocks',
   'pools.trades': 'Trades',
   'pools.pending': 'Bevorstehende',
-  'pools.available': 'Verfügbare',
-  'pools.halted.chain': 'Der Handel in Pools von {chain} Chain(s) ist gestoppt aufgrund von Wartungsarbeiten.'
+  'pools.available': 'Verfügbare'
 }
 
 export default pools