+++ src/renderer/i18n/fr/settings.ts
@@ -2,7 +2,7 @@ import { SettingMessages } from '../types'
 
 const settings: SettingMessages = {
   'setting.title': 'Réglages',
-  'setting.app.title': 'Global settings - FR',
+  'setting.app.title': 'Réglages généraux',
   'setting.wallet.management': 'Gestion du portefeuille',
   'setting.client': 'Client',
   'setting.account.management': 'Gestion du compte',