+++ src/renderer/services/binance/balances.ts
@@ -32,4 +32,6 @@ const balances$: (walletType: WalletType, network: Network) => C.WalletBalancesL
     liveData.map(FP.flow(A.filter(({ asset }) => !assetInBinanceBlacklist(network, asset))))
   )
 
-export { balances$, reloadBalances, reloadBalances$, resetReloadBalances }
+const getBalanceByAddress$ = C.balancesByAddress$(client$, reloadBalances$)
+
+export { balances$, reloadBalances, reloadBalances$, resetReloadBalances, getBalanceByAddress$ }