+++ src/renderer/services/thorchain/const.ts
@@ -11,12 +11,21 @@ export const INITIAL_INTERACT_STATE: InteractState = {
 export const DEFAULT_MIMIR_HALT: MimirHalt = {
   haltThorChain: false,
   haltTrading: false,
-  haltEthTrading: false,
   haltBtcChain: false,
+  haltBtcTrading: false,
   haltEthChain: false,
+  haltEthTrading: false,
   haltBchChain: false,
+  haltBchTrading: false,
   haltLtcChain: false,
-  haltBnbChain: false
+  haltLtcTrading: false,
+  haltBnbChain: false,
+  haltBnbTrading: false,
+  pauseLpBnb: false,
+  pauseLpBch: false,
+  pauseLpBtc: false,
+  pauseLpEth: false,
+  pauseLpLtc: false
 }
 
 export const RESERVE_MODULE_ADDRESS = 'thor1dheycdevq39qlkxs2a6wuuzyn4aqxhve4qxtxt'