+++ src/renderer/services/thorchain/types.ts
@@ -67,7 +67,11 @@ export type NodeDataRD = RD.RemoteData<ApiError, NodeInfo>
 
 export type ThorNodeApiUrlLD = LiveData<ApiError, string>
 
-// https://thornode.thorchain.info/thorchain/mimir
+/**
+ * IO type for mimir endpoints:
+ * mainnet: https://thornode.thorchain.info/thorchain/mimir
+ * testnet: https://testnet.thornode.thorchain.info/thorchain/mimir
+ */
 export const MimirIO = t.type({
   'mimir//MAXIMUMLIQUIDITYRUNE': t.union([t.number, t.undefined]),
   'mimir//POOLCYCLE': t.union([t.number, t.undefined]),
@@ -76,9 +80,18 @@ export const MimirIO = t.type({
   'mimir//HALTETHCHAIN': t.union([t.number, t.undefined]),
   'mimir//HALTETHTRADING': t.union([t.number, t.undefined]),
   'mimir//HALTBTCCHAIN': t.union([t.number, t.undefined]),
+  'mimir//HALTBTCTRADING': t.union([t.number, t.undefined]),
   'mimir//HALTBCHCHAIN': t.union([t.number, t.undefined]),
+  'mimir//HALTBCHTRADING': t.union([t.number, t.undefined]),
   'mimir//HALTLTCCHAIN': t.union([t.number, t.undefined]),
-  'mimir//HALTBNBCHAIN': t.union([t.number, t.undefined])
+  'mimir//HALTLTCTRADING': t.union([t.number, t.undefined]),
+  'mimir//HALTBNBCHAIN': t.union([t.number, t.undefined]),
+  'mimir//HALTBNBTRADING': t.union([t.number, t.undefined]),
+  'mimir//PAUSELPBNB': t.union([t.number, t.undefined]),
+  'mimir//PAUSELPBCH': t.union([t.number, t.undefined]),
+  'mimir//PAUSELPBTC': t.union([t.number, t.undefined]),
+  'mimir//PAUSELPETH': t.union([t.number, t.undefined]),
+  'mimir//PAUSELPLTC': t.union([t.number, t.undefined])
 })
 
 export type Mimir = t.TypeOf<typeof MimirIO>
@@ -86,10 +99,6 @@ export type Mimir = t.TypeOf<typeof MimirIO>
 export type MimirLD = LiveData<Error, Mimir>
 export type MimirRD = RD.RemoteData<Error, Mimir>
 
-export type MimirHalt = MimirHaltChain & {
-  haltEthTrading: boolean
-  haltTrading: boolean
-}
 export type MimirHaltChain = {
   haltThorChain: boolean
   haltBtcChain: boolean
@@ -98,6 +107,25 @@ export type MimirHaltChain = {
   haltLtcChain: boolean
   haltBnbChain: boolean
 }
+export type MimirHaltTrading = {
+  haltTrading: boolean
+  haltBtcTrading: boolean
+  haltEthTrading: boolean
+  haltBchTrading: boolean
+  haltLtcTrading: boolean
+  haltBnbTrading: boolean
+}
+
+export type MimirPauseLP = {
+  pauseLpBnb: boolean
+  pauseLpBch: boolean
+  pauseLpBtc: boolean
+  pauseLpEth: boolean
+  pauseLpLtc: boolean
+}
+
+export type MimirHalt = MimirHaltChain & MimirHaltTrading & MimirPauseLP
+
 export type MimirHaltRD = RD.RemoteData<Error, MimirHalt>
 
 export type GetLiquidityProvidersParams = {