+++ src/renderer/views/pools/PendingPools.tsx
@@ -88,7 +88,9 @@ export const PendingPools: React.FC<PoolsComponentProps> = ({ haltedChains, mimi
 
   const renderBtnPoolsColumn = useCallback(
     (_: string, { pool }: PoolTableRowData) => {
-      const disablePool = PoolHelpers.disableAllActions({ chain: pool.target.chain, haltedChains, mimirHalt })
+      const disablePool =
+        PoolHelpers.disableAllActions({ chain: pool.target.chain, haltedChains, mimirHalt }) ||
+        PoolHelpers.disablePoolActions({ chain: pool.target.chain, haltedChains, mimirHalt })
       return (
         <TableAction>
           <ManageButton asset={pool.target} isTextView={isDesktopView} disabled={disablePool} />