+++ CHANGELOG.md
@@ -1,26 +1,47 @@
+# 0.3.14 (2021-09-04)
+
+## Add
+
+- [Mimir] Support `HALT{chain}TRADING` flags for `BTC`, `LTC`, `BCH`, `BNB` [#1722](https://github.com/thorchain/asgardex-electron/issues/1722)
+- [Mimir] Support `PAUSELP{chain}` flags for `BNB`, `BTC`, `BCH`, `ETH`, `LTC` [#1724](https://github.com/thorchain/asgardex-electron/issues/1724)
+
+## Update
+
+- Updated french translation (01-09-2021) [#1717](https://github.com/thorchain/asgardex-electron/pull/1717)
+
+## Fix
+
+. ETH.RUNE upgrade fails on testnet [#1730](https://github.com/thorchain/asgardex-electron/issues/1730)
+
+- [UpgradeView] Fix rendering hook issues [#1707](https://github.com/thorchain/asgardex-electron/pull/1707)
+
+## Internal
+
+- Update to latest `xchainjs/xchain-binance@5.3.1` [#1725](https://github.com/thorchain/asgardex-electron/issues/1725)
+
 # 0.3.13 (2021-08-28)
 
 ## Add
 
-[Settings] Split settings into application and wallet settings [1575](https://github.com/thorchain/asgardex-electron/issues/1575)
+- [Settings] Split settings into application and wallet settings [1575](https://github.com/thorchain/asgardex-electron/issues/1575)
 
-[Swap] Allow swap to custom recipient [1683](https://github.com/thorchain/asgardex-electron/issues/1683)
+- [Swap] Allow swap to custom recipient [1683](https://github.com/thorchain/asgardex-electron/issues/1683)
 
-[Wallet] Tag fees in txs history of Rune Native [1698](https://github.com/thorchain/asgardex-electron/issues/1698)
+- [Wallet] Tag fees in txs history of Rune Native [1698](https://github.com/thorchain/asgardex-electron/issues/1698)
 
 ## Fix
 
-[Pools] Fix `disableAllActions` helper [1675](https://github.com/thorchain/asgardex-electron/pull/1675)
+- [Pools] Fix `disableAllActions` helper [1675](https://github.com/thorchain/asgardex-electron/pull/1675)
 
-[Upgrade] Add upgrade warning for `BNB.RUNE` [1690](https://github.com/thorchain/asgardex-electron/issues/1690)
+- [Upgrade] Add upgrade warning for `BNB.RUNE` [1690](https://github.com/thorchain/asgardex-electron/issues/1690)
 
-[Swap] Don't send swap transactions to `BNB` accounts with flags > 0 [1611](https://github.com/thorchain/asgardex-electron/issues/1611)
+- [Swap] Don't send swap transactions to `BNB` accounts with flags > 0 [1611](https://github.com/thorchain/asgardex-electron/issues/1611)
 
 ## Internal
 
-Use consistent naming for styles files [1677](https://github.com/thorchain/asgardex-electron/issues/1677)
+- Use consistent naming for styles files [1677](https://github.com/thorchain/asgardex-electron/issues/1677)
 
-SettingsView: Creation of `userAccounts` needs to be more DRY [417](https://github.com/thorchain/asgardex-electron/issues/417)
+- SettingsView: Creation of `userAccounts` needs to be more DRY [417](https://github.com/thorchain/asgardex-electron/issues/417)
 
 # 0.3.12 (2021-08-16)